VoteView 수정

+++ ClazziFirebase/ClazziFirebase/Features/Vote/Views/VoteListView.swift
@@ -5,7 +5,7 @@ import FirebaseAuth
 
 struct VoteListView: View {
     @EnvironmentObject var session: UserSession
-    
+    @StateObject private var voteViewModel = VoteViewModel()
     
     
     // 투표 삭제 관련
@@ -20,30 +20,32 @@ struct VoteListView: View {
         NavigationStack {
             ScrollView {
                 LazyVStack(spacing: 16) {
-//                    ForEach(votes) { vote in
-//                        NavigationLink(destination: VoteView(vote: vote, currentUserID: $currentUserID)) {
-//                            VoteCardView(vote: vote) {
-//                                voteToDelete = vote
-//                                showDeleteAlert = true
-//                            } onEdit: {
-//                                voteToEdit = vote
-//                                isPresentingEdit = true
-//                            }
-//                        }
-//                    }
+                    ForEach(voteViewModel.votes) { vote in
+                        NavigationLink(destination: VoteView(vote: vote) { vote in
+                            voteViewModel.updateVote(vote)
+                        }) {
+                            VoteCardView(vote: vote) {
+                                voteToDelete = vote
+                                showDeleteAlert = true
+                            } onEdit: {
+                                voteToEdit = vote
+                                isPresentingEdit = true
+                            }
+                        }
+                    }
                 }
             }
             .padding()
             .navigationBarTitle("투표 목록 화면")
             .toolbar {
                 // 투표 생성
-//                ToolbarItem(placement: .navigationBarTrailing) {
-//                    NavigationLink(destination: VoteEditorView { newVote in
-//                        modelContext.insert(newVote)
-//                    }) {
-//                        Image(systemName: "plus")
-//                    }
-//                }
+                ToolbarItem(placement: .navigationBarTrailing) {
+                    NavigationLink(destination: VoteEditorView { newVote in
+                        voteViewModel.createVote(newVote)
+                    }) {
+                        Image(systemName: "plus")
+                    }
+                }
                 // 마이페이지
                 ToolbarItem(placement: .navigationBarTrailing) {
                     NavigationLink(destination: MyPageView()) {
@@ -55,26 +57,26 @@ struct VoteListView: View {
             .navigationDestination(isPresented: $isPresentingEdit) {
                 if let vote = voteToEdit {
                     VoteEditorView(vote: vote) { updatedVote in
-                        
+                        voteViewModel.updateVote(updatedVote)
                     }
                 }
             }
             // 삭제 알러트
-//            .alert("투표를 삭제하시겠습니까?", isPresented: $showDeleteAlert) {
-//                Button("삭제", role: .destructive) {
-//                    if let vote = voteToDelete {
-//                        modelContext.delete(vote)
-//                        voteToDelete = nil
-//                    }
-//                }
-//                Button("취소", role: .cancel) {
-//                    voteToDelete = nil // 취소 시 상태 초기화
-//                }
-//            } message: {
-//                if let vote = voteToDelete {
-//                    Text("'\(vote.title)' 투표가 삭제됩니다.")
-//                }
-//            }
+            .alert("투표를 삭제하시겠습니까?", isPresented: $showDeleteAlert) {
+                Button("삭제", role: .destructive) {
+                    if let vote = voteToDelete {
+                        voteViewModel.deleteVote(vote)
+                        voteToDelete = nil
+                    }
+                }
+                Button("취소", role: .cancel) {
+                    voteToDelete = nil // 취소 시 상태 초기화
+                }
+            } message: {
+                if let vote = voteToDelete {
+                    Text("'\(vote.title)' 투표가 삭제됩니다.")
+                }
+            }
         }
     }
 }
@@ -83,7 +85,7 @@ struct VoteCardView: View {
     let vote: Vote
     let onDelete: () -> Void
     let onEdit: () -> Void
-    
+    @EnvironmentObject var session: UserSession
     var body: some View {
         HStack(alignment: .top) {
             VStack(alignment: .leading) {
@@ -95,18 +97,21 @@ struct VoteCardView: View {
                     .foregroundColor(.white)
             }
             Spacer()
-            Button(action: {
-                onEdit()
-            }) {
-                Image(systemName: "pencil")
-                    .foregroundStyle(.white)
+            if let user = session.user, vote.createdBy == user.uid {
+                Button(action: {
+                    onEdit()
+                }) {
+                    Image(systemName: "pencil")
+                        .foregroundStyle(.white)
+                }
+                Button(action: {
+                    onDelete()
+                }) {
+                    Image(systemName: "trash")
+                        .foregroundStyle(.white)
+                }
             }
-            Button(action: {
-                onDelete()
-            }) {
-                Image(systemName: "trash")
-                    .foregroundStyle(.white)
-            }
+            
         }
         .padding(16)
         .frame(maxWidth: .infinity, alignment: .leading)
